+++ src/Fluentify.Tests/Snippets/Classes.AllThreeIgnored.cs
@@ -1,4 +1,4 @@
-﻿namespace Fluentify.Snippets;
+namespace Fluentify.Snippets;
 
 public static partial class Classes
 {
@@ -30,12 +30,6 @@ public static partial class Classes
         "Fluentify.Classes.Testing.AllThreeIgnoredExtensions.With");
 
     private const string AllThreeIgnoredWithExtensionsContent = """
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable enable
-        #endif
-        
-        #pragma warning disable CS8625
-
         namespace Fluentify.Classes.Testing
         {
             using System;
@@ -64,11 +58,5 @@ public static partial class Classes
                 }
             }
         }
-
-        #pragma warning restore CS8625
-        
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable restore
-        #endif
         """;
 }