+++ src/Fluentify.Tests/Snippets/Classes.Boolean.cs
@@ -1,4 +1,4 @@
-﻿namespace Fluentify.Snippets;
+namespace Fluentify.Snippets;
 
 public static partial class Classes
 {
@@ -40,12 +40,6 @@ public static partial class Classes
         "Fluentify.Classes.Testing.BooleanExtensions.With");
 
     private const string BooleanWithAgeExtensionsContent = """
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable enable
-        #endif
-        
-        #pragma warning disable CS8625
-
         namespace Fluentify.Classes.Testing
         {
             using System;
@@ -70,21 +64,9 @@ public static partial class Classes
                 }
             }
         }
-
-        #pragma warning restore CS8625
-        
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable restore
-        #endif
         """;
 
     private const string BooleanWithNameExtensionsContent = """
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable enable
-        #endif
-        
-        #pragma warning disable CS8625
-
         namespace Fluentify.Classes.Testing
         {
             using System;
@@ -109,21 +91,9 @@ public static partial class Classes
                 }
             }
         }
-
-        #pragma warning restore CS8625
-
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable restore
-        #endif
         """;
 
     private const string BooleanIsRetiredExtensionsContent = """
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable enable
-        #endif
-        
-        #pragma warning disable CS8625
-
         namespace Fluentify.Classes.Testing
         {
             using System;
@@ -148,21 +118,9 @@ public static partial class Classes
                 }
             }
         }
-
-        #pragma warning restore CS8625
-        
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable restore
-        #endif
         """;
 
     private const string BooleanWithExtensionsContent = """
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable enable
-        #endif
-        
-        #pragma warning disable CS8625
-
         namespace Fluentify.Classes.Testing
         {
             using System;
@@ -191,11 +149,5 @@ public static partial class Classes
                 }
             }
         }
-
-        #pragma warning restore CS8625
-        
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable restore
-        #endif
         """;
 }