+++ src/Fluentify.Tests/Snippets/Classes.CrossReferenced.cs
@@ -1,4 +1,4 @@
-﻿namespace Fluentify.Snippets;
+namespace Fluentify.Snippets;
 
 public static partial class Classes
 {
@@ -33,12 +33,6 @@ public static partial class Classes
         "Fluentify.Classes.Testing.CrossReferencedExtensions.With");
 
     private const string CrossReferencedWithDescriptionExtensionsContent = """
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable enable
-        #endif
-        
-        #pragma warning disable CS8625
-
         namespace Fluentify.Classes.Testing
         {
             using System;
@@ -62,21 +56,9 @@ public static partial class Classes
                 }
             }
         }
-
-        #pragma warning restore CS8625
-        
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable restore
-        #endif
         """;
 
     private const string CrossReferencedWithSimpleExtensionsContent = """
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable enable
-        #endif
-        
-        #pragma warning disable CS8625
-
         namespace Fluentify.Classes.Testing
         {
             using System;
@@ -134,21 +116,9 @@ public static partial class Classes
                 }
             }
         }
-
-        #pragma warning restore CS8625
-        
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable restore
-        #endif
         """;
 
     private const string CrossReferencedWithExtensionsContent = """
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable enable
-        #endif
-        
-        #pragma warning disable CS8625
-
         namespace Fluentify.Classes.Testing
         {
             using System;
@@ -174,11 +144,5 @@ public static partial class Classes
                 }
             }
         }
-
-        #pragma warning restore CS8625
-        
-        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        #nullable restore
-        #endif
         """;
 }