+++ src/Fluentify/RecordGenerator.cs
@@ -38,16 +38,6 @@ public sealed class RecordGenerator
         }
     }
 
-    /// <inheritdoc/>
-    private protected override string Wrap(string content)
-    {
-        return $"""
-            #nullable enable
-            {content}
-            #nullable restore
-            """;
-    }
-
     private static Source GenerateConstructor(Subject subject)
     {
         string content = subject.GetConstructor();